+++ examples/starter/.storybook/preview.ts
@@ -1,4 +1,4 @@
-import type { Preview } from '@nuxtjs/storybook'
+import type { Preview } from 's8n11c-nuxtjs-storybook'
 
 const preview: Preview = {
   parameters: {