+++ examples/tailwind/nuxt.config.ts
@@ -1,7 +1,7 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   devtools: { enabled: true },
-  modules: ['@nuxtjs/storybook', '@nuxtjs/tailwindcss'],
+  modules: ['s8n11c-nuxtjs-storybook', '@nuxtjs/tailwindcss'],
 
   compatibilityDate: '2024-08-03',
 })