+++ playground/.storybook/main.ts
@@ -1,4 +1,4 @@
-import type { StorybookConfig } from '@nuxtjs/storybook'
+import type { StorybookConfig } from 's8n11c-nuxtjs-storybook'
 import Inspect from 'vite-plugin-inspect'
 
 const config: StorybookConfig = {
@@ -8,7 +8,7 @@ const config: StorybookConfig = {
   ],
   addons: ['@storybook/addon-docs', '@chromatic-com/storybook'],
   framework: {
-    name: '@storybook-vue/nuxt',
+    name: 's8n11c-storybook-vue-nuxt',
     options: {},
   },
   viteFinal: (config) => {